RaspberryPi/requirements.txt
tensorflow==1.7.0
numpy==1.14.2
paho_mqtt==1.3.1
